Bump to Terraform 0.12-beta1
This updates the Terraform plugin provider to match the new version.
This uses the version 5 of plugin interface instead of 4 from Terraform
0.11.x.

There are a couple of transitive dependencies that require Bazaar to
download and compile. This make compiling harder on Windows (where it
was tested) and requires an extra dependency to exist on computers.

This commit replaces such dependencies with alternative project paths
which either use Git or HTTPS to keep compilation simple on other
systems.